backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
pytz
tzdata
PyJWT
bcrypt
python-jose[cryptography]
